// frontend/src/api/generated/models/torrent-summary-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * My API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface TorrentSummaryDTO
 */
export interface TorrentSummaryDTO {
    /**
     * 
     * @type {string}
     * @memberof TorrentSummaryDTO
     */
    'hash'?: string;
    /**
     * 
     * @type {string}
     * @memberof TorrentSummaryDTO
     */
    'name'?: string;
    /**
     * 
     * @type {string}
     * @memberof TorrentSummaryDTO
     */
    'status'?: TorrentSummaryDTOStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof TorrentSummaryDTO
     */
    'state'?: string;
    /**
     * 
     * @type {number}
     * @memberof TorrentSummaryDTO
     */
    'progress'?: number;
    /**
     * 
     * @type {number}
     * @memberof TorrentSummaryDTO
     */
    'size'?: number;
    /**
     * 
     * @type {number}
     * @memberof TorrentSummaryDTO
     */
    'downloaded'?: number;
    /**
     * 
     * @type {number}
     * @memberof TorrentSummaryDTO
     */
    'uploaded'?: number;
    /**
     * 
     * @type {number}
     * @memberof TorrentSummaryDTO
     */
    'downloadSpeed'?: number;
    /**
     * 
     * @type {number}
     * @memberof TorrentSummaryDTO
     */
    'uploadSpeed'?: number;
    /**
     * 
     * @type {number}
     * @memberof TorrentSummaryDTO
     */
    'peers'?: number;
    /**
     * 
     * @type {number}
     * @memberof TorrentSummaryDTO
     */
    'seeds'?: number;
    /**
     * 
     * @type {number}
     * @memberof TorrentSummaryDTO
     */
    'etaSeconds'?: number;
    /**
     * 
     * @type {string}
     * @memberof TorrentSummaryDTO
     */
    'category'?: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof TorrentSummaryDTO
     */
    'tags'?: Array<string>;
    /**
     * 
     * @type {number}
     * @memberof TorrentSummaryDTO
     */
    'ratio'?: number;
}

export const TorrentSummaryDTOStatusEnum = {
    Downloading: 'DOWNLOADING',
    Seeding: 'SEEDING',
    Paused: 'PAUSED',
    Completed: 'COMPLETED'
} as const;

export type TorrentSummaryDTOStatusEnum = typeof TorrentSummaryDTOStatusEnum[keyof typeof TorrentSummaryDTOStatusEnum];